# backend/API/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ErrorResponse:
    properties:
      code:
        description: |-
          The HTTP status code of the response.
          example: 400
        type: integer
      error:
        description: |-
          The standard HTTP status text for the code.
          example: Bad Request
        type: string
      message:
        description: |-
          A user-friendly message describing the specific issue.
          example: Invalid request body or validation failed.
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the API documentation for the social-network application.
  title: Social-network Application API
  version: "1.0"
paths:
...
  /api/v1/comments/{id}:
    delete:
      description: Performs a soft-delete on a comment. Only the author can perform
        this action.
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'status: deleted'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Comment not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Delete a comment
      tags:
      - Comments
    put:
      consumes:
      - application/json
      description: Updates the text content of an existing comment. Only the author
        can perform this action.
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'status: updated'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - not the owner
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Comment not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Edit a comment
      tags:
      - Comments
  /api/v1/comments/create:
    post:
      consumes:
      - multipart/form-data
      description: Adds a new comment to a specific post. Optionally upload images.
        Triggers a real-time notification for the post owner via WebSocket.
      parameters:
      - description: Post ID
        in: formData
        name: post_id
        required: true
        type: string
      - description: Comment content
        in: formData
        name: content
        required: true
        type: string
      - description: Images to attach (multiple allowed)
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Comment with images_uploaded count
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Missing PostID or Content
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Create a comment
      tags:
      - Comments
...
swagger: "2.0"